Ai/requirements.txt
fastapi[all]
openai
python-dotenv
